+++ src/opcode.h
@@ -107,7 +107,9 @@ namespace ydsh {
   OP(ADD_EXPANDING, 2, 0)                                                                          \
   OP(CALL_CMD, 0, -1)                                                                              \
   OP(CALL_CMD_NOFORK, 0, -1)                                                                       \
+  OP(CALL_CMD_SILENT, 0, -1)                                                                       \
   OP(CALL_UDC, 2, -1)                                                                              \
+  OP(CALL_UDC_SILENT, 2, -1)                                                                       \
   OP(CALL_CMD_COMMON, 0, -1)                                                                       \
   OP(BUILTIN_CMD, 0, 1)                                                                            \
   OP(BUILTIN_EVAL, 0, 1)                                                                           \